+++ Spring/project_01/src/main/java/org/moon/service/ReplyService.java
@@ -13,7 +13,7 @@ public interface ReplyService {
 	
 	public ReplyVO replyRead(Long rno);
 	
-	public int replyDelete(int rno);
+	public int replyDelete(Long rno);
 	
 	public int replyModify(ReplyVO vo);
